fix: resolve CI failures - file line limits, code duplication, and formatting
- Extract displayModelUsage, displayCostComparison, createEmptySubSessionUsage,
  and accumulateModelUsage from claude.lib.mjs to claude.budget-stats.lib.mjs
  to keep all files under 1500 line limit
- Pre-compute budgetStats string once in attachLogToGitHub instead of 4 times
- Refactor test-budget-stats.mjs to import actual module and use shared fixtures,
  reducing code duplication from 11.15% to 10.9% (under 11% threshold)
